+++ valkey/specs/requirements/aws/main.tf
@@ -82,17 +82,30 @@ resource "aws_iam_policy" "nullplatform_valkey_state" {
 
   policy = jsonencode({
     Version = "2012-10-17"
-    Statement = [{
-      Sid    = "ManageStateBuckets"
-      Effect = "Allow"
-      Action = ["s3:*"]
-      Resource = [
-        "arn:aws:s3:::np-service-*",
-        "arn:aws:s3:::np-service-*/*",
-        "arn:aws:s3:::null-service-provisioning-kwik-e-mart-main",
-        "arn:aws:s3:::null-service-provisioning-kwik-e-mart-main/*",
-      ]
-    }]
+    Statement = [
+      {
+        Sid    = "ListStateBucket"
+        Effect = "Allow"
+        Action = [
+          "s3:ListBucket",
+          "s3:ListBucketVersions",
+          "s3:GetBucketLocation",
+        ]
+        Resource = "arn:aws:s3:::${var.state_bucket_name}"
+      },
+      {
+        Sid    = "ManageStateObjects"
+        Effect = "Allow"
+        Action = [
+          "s3:GetObject",
+          "s3:GetObjectVersion",
+          "s3:PutObject",
+          "s3:DeleteObject",
+          "s3:DeleteObjectVersion",
+        ]
+        Resource = "arn:aws:s3:::${var.state_bucket_name}/*"
+      },
+    ]
   })
 
   tags = local.iam_default_tags